Add support for re-using sessions
Previously, multiple logins would produce new sessions on each request. Now sessions will be reused on each login attempt until the session is either destroyed (via log out) or has expired. In addition, API key access was also added on a per-user basis. Creating a session for an API auth request is similar to login, and will reuse the session on subsequent requests. API tokens/permissions are scoped to the user the key is issued to.